requirements.txt
matplotlib>=2.0.0
pandas>=1.0.0
PyYAML>=3.0.1
PyYAML>=3.0.1
redis>=3.0.0
seaborn>0.0.0
SQLAlchemy>1.0.0
streamlit>1.0.0
streamlit_authenticator>=0.1.0
streamlit_extras>=0.3.0
streamlit_option_menu>=0.3.0
